fix: fixed up download handler to improve error handling
