+++ src/core/server/app/handlers/api/index.ts
@@ -4,3 +4,4 @@ export * from "./graphql";
 export * from "./health";
 export * from "./install";
 export * from "./version";
+export * from "./user";
