+++ src/core/server/app/handlers/api/user/index.ts
@@ -1,0 +1,1 @@
+export * from "./download";
