+++ src/core/server/app/router/api/account.ts
@@ -3,12 +3,11 @@ import express from "express";
 
 import { AppOptions } from "coral-server/app";
 import {
-  adminDownloadHandler,
+  accountDownloadCheckHandler,
+  accountDownloadHandler,
   confirmCheckHandler,
   confirmHandler,
   confirmRequestHandler,
-  downloadCheckHandler,
-  downloadHandler,
   inviteCheckHandler,
   inviteHandler,
 } from "coral-server/app/handlers";
@@ -34,14 +33,13 @@ export function createNewAccountRouter(
   router.get("/invite", inviteCheckHandler(app));
   router.put("/invite", jsonMiddleware, inviteHandler(app));
 
-  router.get("/downloadcheck", downloadCheckHandler(app));
-  router.get("/download", adminDownloadHandler(app));
+  router.get("/download", accountDownloadCheckHandler(app));
   router.post(
     "/download",
     bodyParser.urlencoded({
       extended: true,
     }),
-    downloadHandler(app)
+    accountDownloadHandler(app)
   );
 
   return router;
